+++ src/ResourceManager/ApiManagement/Commands.ApiManagement.ServiceManagement/Commands/NewAzureApiManagementUser.cs
@@ -17,6 +17,8 @@ namespace Microsoft.Azure.Commands.ApiManagement.ServiceManagement.Commands
     using Microsoft.Azure.Commands.ApiManagement.ServiceManagement.Models;
     using System;
     using System.Management.Automation;
+    using System.Security;
+    using WindowsAzure.Commands.Common;
 
     [Cmdlet(VerbsCommon.New, Constants.ApiManagementUser)]
     [OutputType(typeof(PsApiManagementUser))]
@@ -61,7 +63,7 @@ namespace Microsoft.Azure.Commands.ApiManagement.ServiceManagement.Commands
             Mandatory = true,
             HelpMessage = "User password. This parameter is required.")]
         [ValidateNotNullOrEmpty]
-        public String Password { get; set; }
+        public SecureString Password { get; set; }
 
         [Parameter(
             ValueFromPipelineByPropertyName = true,
@@ -78,8 +80,8 @@ namespace Microsoft.Azure.Commands.ApiManagement.ServiceManagement.Commands
         public override void ExecuteApiManagementCmdlet()
         {
             string userId = UserId ?? Guid.NewGuid().ToString("N");
-
-            var user = Client.UserCreate(Context, userId, FirstName, LastName, Password, Email, State, Note);
+            
+            var user = Client.UserCreate(Context, userId, FirstName, LastName, Password.ConvertToString(), Email, State, Note);
 
             WriteObject(user);
         }