+++ src/ResourceManager/Compute/Commands.Compute.Test/ScenarioTests/VirtualMachineTests.ps1
@@ -169,6 +169,8 @@ function Test-VirtualMachine
         Update-AzureRmVM -ResourceGroupName $rgname -VM $p;
 
         $vm2 = Get-AzureRmVM -Name $vmname -ResourceGroupName $rgname;
+        Assert-AreEqual $null $vm2.Zones;
+
         Assert-AreEqual $vm2.NetworkProfile.NetworkInterfaces.Count 1;
         Assert-AreEqual $vm2.NetworkProfile.NetworkInterfaces[0].Id $nicId;
 
@@ -2830,7 +2832,6 @@ function Test-VirtualMachineGetStatus
         $a = $vms | Out-String;
         Write-Verbose($a);
         Assert-True {$a.Contains("PowerState");}
-        Assert-True {$a.Contains("Maintenance");}
 
         $vms = Get-AzureRmVM -Status;
         $a = $vms | Out-String;
@@ -3121,7 +3122,7 @@ function Test-VirtualMachineIdentity
         $computerName = 'test';
         $vhdContainer = "https://$stoname.blob.core.windows.net/test";
 
-        $p = New-AzureRmVMConfig -VMName $vmname -VMSize $vmsize -IdentityType "SystemAssigned" `
+        $p = New-AzureRmVMConfig -VMName $vmname -VMSize $vmsize -AssignIdentity `
              | Add-AzureRmVMNetworkInterface -Id $nicId -Primary `
              | Set-AzureRmVMOSDisk -Name $osDiskName -VhdUri $osDiskVhdUri -Caching $osDiskCaching -CreateOption FromImage `
              | Set-AzureRmVMOperatingSystem -Windows -ComputerName $computerName -Credential $cred;
@@ -3218,7 +3219,7 @@ function Test-VirtualMachineIdentityUpdate
         $vms_output = $vms | Out-String;
         Write-Verbose($vms_output);
 
-        $st = $vm1 | Update-AzureRmVM -IdentityType "SystemAssigned"
+        $st = $vm1 | Update-AzureRmVM -AssignIdentity;
 
         # Get VM
         $vm1 = Get-AzureRmVM -Name $vmname -ResourceGroupName $rgname -DisplayHint "Expand";