+++ src/ResourceManager/EventGrid/Commands.EventGrid.Test/ScenarioTests/EventGridController.cs
@@ -1,0 +1,166 @@
+﻿// ----------------------------------------------------------------------------------
+//
+// Copyright Microsoft Corporation
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ----------------------------------------------------------------------------------
+
+using Microsoft.Azure.Commands.Common.Authentication;
+using Microsoft.Azure.Management.Resources;
+using Microsoft.Azure.Subscriptions;
+using Microsoft.Azure.Test.HttpRecorder;
+using Microsoft.Rest.ClientRuntime.Azure.TestFramework;
+using Microsoft.WindowsAzure.Commands.ScenarioTest;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using Microsoft.Azure.Management.EventGrid;
+using LegacyTest = Microsoft.Azure.Test;
+using TestEnvironmentFactory = Microsoft.Rest.ClientRuntime.Azure.TestFramework.TestEnvironmentFactory;
+using TestUtilities = Microsoft.Rest.ClientRuntime.Azure.TestFramework.TestUtilities;
+using Microsoft.Azure.Management.EventHub;
+
+namespace Microsoft.Azure.Commands.EventGrid.Test.ScenarioTests
+{
+    public class EventGridController
+    {
+        private LegacyTest.CSMTestEnvironmentFactory csmTestFactory;
+        private EnvironmentSetupHelper helper;
+        private const string TenantIdKey = "TenantId";
+        private const string DomainKey = "Domain";
+        private const string AuthorizationApiVersion = "2014-07-01-preview";
+
+        public ResourceManagementClient ResourceManagementClient { get; private set; }
+
+        public EventGridManagementClient EventGridManagementClient { get; private set; }
+
+        public EventHubManagementClient EventHubClient { get; private set; }
+
+        public string UserDomain { get; private set; }
+
+        public static EventGridController NewInstance
+        {
+            get
+            {
+                return new EventGridController();
+            }
+        }
+
+
+        public EventGridController()
+        {
+            helper = new EnvironmentSetupHelper();
+        }
+
+        public void RunPsTest(params string[] scripts)
+        {
+            var callingClassType = TestUtilities.GetCallingClass(2);
+            var mockName = TestUtilities.GetCurrentMethodName(2);
+
+            this.RunPsTestWorkflow(
+                () => scripts,
+                // no custom initializer
+                null,
+                // no custom cleanup
+                null,
+                callingClassType,
+                mockName);
+        }
+
+        public void RunPsTestWorkflow(
+            Func<string[]> scriptBuilder,
+            Action<LegacyTest.CSMTestEnvironmentFactory> initialize,
+            Action cleanup,
+            string callingClassType,
+            string mockName)
+        {
+            Dictionary<string, string> d = new Dictionary<string, string>();
+            d.Add("Microsoft.Resources", null);
+            d.Add("Microsoft.Features", null);
+            d.Add("Microsoft.Authorization", null);
+            var providersToIgnore = new Dictionary<string, string>();
+            providersToIgnore.Add("Microsoft.Azure.Management.Resources.ResourceManagementClient", "2016-02-01");
+            HttpMockServer.Matcher = new PermissiveRecordMatcherWithApiExclusion(true, d, providersToIgnore);
+
+            HttpMockServer.RecordsDirectory = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "SessionRecords");
+            using (MockContext context = MockContext.Start(callingClassType, mockName))
+            {
+                this.csmTestFactory = new LegacyTest.CSMTestEnvironmentFactory();
+                if (initialize != null)
+                {
+                    initialize(this.csmTestFactory);
+                }
+                SetupManagementClients(context);
+                helper.SetupEnvironment(AzureModule.AzureResourceManager);
+
+                string callingClassName = callingClassType
+                                        .Split(new[] { "." }, StringSplitOptions.RemoveEmptyEntries)
+                                        .Last();
+                helper.SetupModules(AzureModule.AzureResourceManager,
+                    "ScenarioTests\\Common.ps1",
+                    "ScenarioTests\\" + callingClassName + ".ps1",
+                    helper.RMProfileModule,
+                    helper.RMResourceModule,
+                    helper.GetRMModulePath(@"AzureRM.EventHub.psd1"),
+                    helper.GetRMModulePath(@"AzureRM.EventGrid.psd1"),
+                    "AzureRM.Resources.ps1");
+
+                try
+                {
+                    if (scriptBuilder != null)
+                    {
+                        var psScripts = scriptBuilder();
+
+                        if (psScripts != null)
+                        {
+                            helper.RunPowerShellTest(psScripts);
+                        }
+                    }
+                }
+                finally
+                {
+                    if (cleanup != null)
+                    {
+                        cleanup();
+                    }
+                }
+            }
+        }
+
+        private void SetupManagementClients(MockContext context)
+        {
+            ResourceManagementClient = GetResourceManagementClient();
+            EventGridManagementClient = GetEventGridManagementClient(context);
+            EventHubClient = GetEHClient(context);
+            helper.SetupManagementClients(EventHubClient, ResourceManagementClient, EventGridManagementClient);
+        }
+
+        private ResourceManagementClient GetResourceManagementClient()
+        {
+            return LegacyTest.TestBase.GetServiceClient<ResourceManagementClient>(this.csmTestFactory);
+        }
+
+        private EventGridManagementClient GetEventGridManagementClient(MockContext context)
+        {
+            return context.GetServiceClient<EventGridManagementClient>(TestEnvironmentFactory.GetTestEnvironment());
+        }
+
+        private EventHubManagementClient GetEHClient(MockContext context)
+        {
+            return context.GetServiceClient<EventHubManagementClient>(TestEnvironmentFactory.GetTestEnvironment());
+        }
+
+        private SubscriptionClient GetSubscriptionClient()
+        {
+            return LegacyTest.TestBase.GetServiceClient<SubscriptionClient>(this.csmTestFactory);
+        }
+    }
+}