+++ src/ResourceManager/Network/Commands.Network/ApplicationGateway/SslCertificate/AzureApplicationGatewaySslCertificateBase.cs
@@ -38,6 +38,7 @@ namespace Microsoft.Azure.Commands.Network
                Mandatory = true,
                HelpMessage = "Certificate password")]
         [ValidateNotNullOrEmpty]
+        [Obsolete("(Get/Set/New)-AzureRmApplicationGatewaySslCertificate: The parameter \"Password\" is being changed from a string to a SecureString in an upcoming breaking change release.")]
         public string Password { get; set; }
 
         public PSApplicationGatewaySslCertificate NewObject()
@@ -46,7 +47,9 @@ namespace Microsoft.Azure.Commands.Network
 
             sslCertificate.Name = this.Name;
             sslCertificate.Data = Convert.ToBase64String(File.ReadAllBytes(CertificateFile));
+#pragma warning disable 0618
             sslCertificate.Password = this.Password;
+#pragma warning restore 0618
             sslCertificate.Id =
                 ApplicationGatewayChildResourceHelper.GetResourceNotSetId(
                     this.NetworkClient.NetworkManagementClient.SubscriptionId,