+++ src/ResourceManager/Resources/Commands.Resources/ActiveDirectory/NewAzureADUserCommand.cs
@@ -14,6 +14,7 @@
 
 using Microsoft.Azure.Graph.RBAC.Version1_6.ActiveDirectory;
 using Microsoft.Azure.Graph.RBAC.Version1_6.Models;
+using System;
 using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.ActiveDirectory
@@ -34,6 +35,7 @@ namespace Microsoft.Azure.Commands.ActiveDirectory
 
         [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = "Password for the user.")]
         [ValidateNotNullOrEmpty]
+        [Obsolete("New-AzureRmADUser: The parameter \"Password\" is being changed from a string to a SecureString in an upcoming breaking change release.")]
         public string Password { get; set; }
 
         [Parameter(Mandatory = false, ValueFromPipelineByPropertyName = true, HelpMessage = "ImmutableId - to be specified only if you are using a federated domain for the user's user principal name (upn) property.")]
@@ -52,7 +54,9 @@ namespace Microsoft.Azure.Commands.ActiveDirectory
                 DisplayName = DisplayName,
                 PasswordProfile = new PasswordProfile
                 {
+#pragma warning disable 0618
                     Password = Password,
+#pragma warning restore 0618
                     ForceChangePasswordNextLogin = ForceChangePasswordNextLogin.IsPresent ? true : false
                 },
                 UserPrincipalName = UserPrincipalName