+++ src/ResourceManager/ServiceBus/Commands.ServiceBus/Cmdlets/Subscription/NewAzureServiceBusSubscription.cs
@@ -20,7 +20,7 @@ using System.Xml;
 namespace Microsoft.Azure.Commands.ServiceBus.Commands.Subscription
 {
     /// <summary>
-    /// 'New-AzureRmServiceBusSubscription' Cmdlet creates a new EventHub
+    /// 'New-AzureRmServiceBusSubscription' Cmdlet creates a new Subscription
     /// </summary>
     [Cmdlet(VerbsCommon.New, ServicebusSubscriptionVerb, SupportsShouldProcess = true), OutputType(typeof(SubscriptionAttributes))]
     public class NewAzureRmServiceBusSubscription : AzureServiceBusCmdletBase
@@ -71,14 +71,6 @@ namespace Microsoft.Azure.Commands.ServiceBus.Commands.Subscription
 
         [Parameter(Mandatory = false,
             ValueFromPipelineByPropertyName = true,
-            HelpMessage = "DeadLetteringOnFilterEvaluationExceptions - Value that indicates if a subscription has dead letter support when a message expires.")]
-        [ValidateSet("TRUE", "FALSE",
-            IgnoreCase = true)]
-        [ValidateNotNullOrEmpty]
-        public bool? DeadLetteringOnFilterEvaluationExceptions { get; set; }
-
-        [Parameter(Mandatory = false,
-            ValueFromPipelineByPropertyName = true,
             HelpMessage = "Dead Lettering On Message Expiration")]
         [ValidateSet("TRUE", "FALSE",
             IgnoreCase = true)]
@@ -91,15 +83,7 @@ namespace Microsoft.Azure.Commands.ServiceBus.Commands.Subscription
         [ValidateSet("TRUE", "FALSE",
             IgnoreCase = true)]
         [ValidateNotNullOrEmpty]
-        public bool? EnableBatchedOperations { get; set; }        
-
-        [Parameter(Mandatory = false,
-            ValueFromPipelineByPropertyName = true,
-            HelpMessage = "IsReadOnly - Value that indicates whether the entity description is read-only.")]
-        [ValidateSet("TRUE", "FALSE",
-            IgnoreCase = true)]
-        [ValidateNotNullOrEmpty]
-        public bool? IsReadOnly { get; set; }
+        public bool? EnableBatchedOperations { get; set; }
 
         [Parameter(Mandatory = false,
             ValueFromPipelineByPropertyName = true,
@@ -152,15 +136,7 @@ namespace Microsoft.Azure.Commands.ServiceBus.Commands.Subscription
 
             if (RequiresSession != null)
                 subAttributes.RequiresSession = RequiresSession;
-
-#pragma warning disable 612, 618
-            if (IsReadOnly != null)
-                subAttributes.IsReadOnly = IsReadOnly;
-            if (DeadLetteringOnFilterEvaluationExceptions != null)
-                subAttributes.DeadLetteringOnFilterEvaluationExceptions = DeadLetteringOnFilterEvaluationExceptions;
-            subAttributes.Location = getNamespaceLoc.Location;
-#pragma warning restore 612, 618
-
+            
             if (ShouldProcess(target: Name, action: string.Format(Resources.CreateSubscription, Name, Topic,Namespace)))
             {
                 WriteObject(Client.CreateUpdateSubscription(ResourceGroupName, Namespace, Topic, Name, subAttributes));