+++ src/ResourceManager/Sql/Commands.Sql/Auditing/Cmdlet/GetAzureSqlDatabaseAuditingPolicy.cs
@@ -22,7 +22,7 @@ namespace Microsoft.Azure.Commands.Sql.Auditing.Cmdlet
     /// Returns the auditing policy of a specific database.
     /// </summary>
     [Cmdlet(VerbsCommon.Get, "AzureRmSqlDatabaseAuditingPolicy", SupportsShouldProcess = true), OutputType(typeof (AuditingPolicyModel))]
-    [Obsolete("Get-AzureRmSqlDatabaseAuditingPolicy is obsolete. It will be removed in a future release. Please use the Get-AzureSqlDatabaseAuditing cmdlet instead.", false)]
+    [Obsolete("Note that Table auditing is deprecated and this command will be removed in a future release. Please use the 'Get-AzureRmSqlDatabaseAuditing' command to get Blob auditing settings.", false)]
     public class GetAzureSqlDatabaseAuditingPolicy : SqlDatabaseAuditingCmdletBase
     {
         /// <summary>